requirements.txt
aiogram-dialog==2.1.0
asyncpg==0.29.0
python-dotenv==1.0.1
python-string-utils==1.0.0
setuptools==68.2.0
SQLAlchemy==2.0.29
pwnedpasswords==2.0.0